pexels_service/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
pydantic==2.11.2
python-dotenv==1.1.0
httpx==0.28.1